requirements.txt
paho-mqtt
pika
elasticsearch==7.15.1
pytz
tzlocal
python-dateutil
pyyaml
